+++ src/chrome/local-runtime.ts
@@ -34,6 +34,7 @@ import {
   mouseup,
   focus,
   clearInput,
+  selectFiles,
 } from '../util'
 
 export default class LocalRuntime {
@@ -103,6 +104,8 @@ export default class LocalRuntime {
         return this.focus(command.selector)
       case 'clearInput':
         return this.clearInput(command.selector)
+      case 'selectFiles':
+        return this.selectFiles(command.selector, command.files)
       default:
         throw new Error(`No such command: ${JSON.stringify(command)}`)
     }
@@ -399,6 +402,21 @@ export default class LocalRuntime {
     }
     await clearInput(this.client, selector)
     this.log(`${selector} cleared`)
+  }
+
+  async selectFiles(selector: string, files: string[]): Promise<void> {
+    if (this.chromelessOptions.implicitWait) {
+      this.log(`selectFiles(): Waiting for ${selector}`)
+      await waitForNode(this.client, selector, this.chromelessOptions.waitTimeout)
+    }
+
+    const exists = await nodeExists(this.client, selector)
+    if (!exists) {
+      throw new Error(`selectFiles(): node for selector ${selector} doesn't exist`)
+    }
+
+    await selectFiles(this.client, selector, files)
+    this.log(`Selected files ${files}`)
   }
 
   private log(msg: string): void {